refactor: move internal plumbing types out of public types.ts
AudioBufferQueueItem and StreamHandler are internal glue between
SendspinCore and the audio scheduler / protocol handler. Exporting them
via index.ts's export * shrinks future refactor freedom and leaks
implementation detail. Move to src/internal-types.ts, which is not
re-exported.
